Fix static analysis issues

+++ source/CoordinateConversion/CoordinateConversionLibrary/Helpers/ExportCSV.cs
@@ -132,7 +132,7 @@ namespace Jitbit.Utils
             if (value is INullable && ((INullable)value).IsNull) return "";
             if (value is DateTime)
             {
-                if (((DateTime)value).TimeOfDay.TotalSeconds == 0)
+                if ((int)((DateTime)value).TimeOfDay.TotalSeconds == 0)
                     return ((DateTime)value).ToString("yyyy-MM-dd");
                 return ((DateTime)value).ToString("yyyy-MM-dd HH:mm:ss");
             }
